requirements.txt
streamlit
openai
python-dotenv
requests
pyperclip
configparser
modelscope
funasr              
torch      
torchvision       
torchaudio          
tensorflow          
logging
librosa
soundfile
typeguard
addict
datasets == 2.16
simplejson
sortedcontainers
